Simplify not importing InputFile
