+++ proxied.d.ts
@@ -4,13 +4,13 @@ import {
   ReplyKeyboardMarkup,
   ReplyKeyboardRemove,
 } from "./callback";
-import { File } from "./file";
 import { InlineKeyboardMarkup, InlineQueryResult } from "./inline";
 import {
   BotCommand,
   ChatFromGetChat,
   ChatMember,
   ChatPermissions,
+  File,
   UserFromGetMe,
   UserProfilePhotos,
   WebhookInfo,
@@ -383,8 +383,8 @@ export interface Typegram<F> {
       duration?: Integer;
       /** Video width and height, i.e. diameter of the video message */
       length?: Integer;
-      /** | String Thumbnail of the file sent; can be ignored if thumbnail generation for the file is supported server-side. The thumbnail should be in JPEG format and less than 200 kB in size. A thumbnail's width and height should not exceed 320. Ignored if the file is not uploaded using multipart/form-data. Thumbnails can't be reused and can be only uploaded as a new file, so you can pass “attach://<file_attach_name>” if the thumbnail was uploaded using multipart/form-data under <file_attach_name>. */
-      thumb?: F;
+      /**  Thumbnail of the file sent; can be ignored if thumbnail generation for the file is supported server-side. The thumbnail should be in JPEG format and less than 200 kB in size. A thumbnail's width and height should not exceed 320. Ignored if the file is not uploaded using multipart/form-data. Thumbnails can't be reused and can be only uploaded as a new file, so you can pass “attach://<file_attach_name>” if the thumbnail was uploaded using multipart/form-data under <file_attach_name>. */
+      thumb?: F | String;
       /** Sends the message silently. Users will receive a notification with no sound. */
       disable_notification?: Boolean;
       /** If the message is a reply, ID of the original message */
